requirements.txt
beautifulsoup4
soupsieve
statistics
requests
